requirements.txt
Pillow==2.8.1
tempdir==0.6
